very primitive split of builder class
